% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{client}
\alias{client}
\title{LakehouseClient Package}
\usage{
client(url)
}
\arguments{
\item{url}{String containing the lakehouse-api url}

\item{email}{A string containing the user email}

\item{password}{A string containing the user password}

\item{catalog_file_id}{A string with the catalog id for the file record to be downloaded}

\item{output_file_dir}{A string containing the output file directory}

\item{storage_type}{A string containing the storage type ("gcs" for google cloud storage, "s3" for amazon s3, "hdfs" for hadoop hdfs)}

\item{collection_name}{A string containing the new collection's name}

\item{namenode_address}{An optional string containing the ip address or URL to the Hadoop namenode}

\item{bucket_name}{An optional string containing the cloud bucket name}

\item{collection_description}{An optional string containing additional description for this collection}

\item{secret}{Optional boolean value indicating whether the collection will be secret (TRUE or FALSE)}

\item{catalogue_file_id}{A string with the catalog id for the file record to be downloaded}

\item{include_raw}{A Boolean value indicating if the raw files should be included in the response}

\item{include_processed}{A Boolean value indicating if the processed files should be included in the response}

\item{include_curated}{A Boolean value indicating if the curated files should be included in the response}

\item{sort_by_key}{A string containing a valid catalog key to be user for sorting the response}

\item{sort_desc}{A boolean indicating if the reponse list should be sorted descending}

\item{df:}{The R dataframe contaiing the data}

\item{df_name:}{the dataframe name only, without the extension. By default the dataframe will be stored as a CSV file}

\item{dataframe_description}{[Optional]: Additional description for the file}

\item{dataframe_version}{[Optional, default 1]: The version of this dataframe in the system}

\item{local_file_path:}{the local path to the file to be uploaded}

\item{final_file_name:}{the output file name in the storage}

\item{collection_catalog_id:}{the collection identifiyer, from the collection catalog, where the file will be placed in}

\item{file_category:}{the file class must indicate if the file is 'structured' or 'unstructured', by default the file is set to be 'unstructured'. Structured files can be Columnar or document files such as csv, tsv, excel, json, parquet.}

\item{file_description}{[Optional]: Additional description for the file}

\item{file_version}{[Optional, default 1]: The version of the file you are uploading for version control}

\item{file_size}{[Optional, default 0]: The size of the file you are uploading in bytes, the sizen will be taken by default from your system}

\item{public}{[Optional, default False]: The visibility of the dataframe, if public all users can see in the catalog}

\item{processing_level}{[Optional]: A string containing the processing level of the file to be uploaded, e.g., ["raw", "processed", "curated"]}

\item{keyword}{A string containing the keyword to search for, the search will match the file names to the keyword}

\item{...}{Query strings containing search terms in KEY[OPERATOR]VALUE format}
}
\value{
A string with the path to the output file

A dataset containing the collection's information in the catalog

A dataframe representing the file selected

Returns an R dataframe with the storage buckets in the system

Returns a json-formatted string with the storage buckets in the system

A R dataframe cotaining the collection records

A R dataframe cotaining the file records

A json-formatted string cotaining the file records

Returns a dataframe containing the catalog record for the inserted file

A list of collections in the given output_format

A list of files in the given output_format

The search results in the requested format. Returns an empty list if no results found or if error occurs.

The search results in the requested format. Returns an empty list if no results found or if error occurs.
}
\description{
Search files on the catalogue based on the given filters using query strings.

Search files on the catalogue based on the given filters using query strings.
}
\section{Query String Structure}{

The query string should follow the pattern: \code{KEY[OPERATOR]VALUE}\cr
Supported operators:
\itemize{
  \item \code{=} - Equal to
  \item \code{>} - Greater than
  \item \code{<} - Less than
  \item \code{>=} - Greater than or equal to
  \item \code{<=} - Less than or equal to
  \item \code{*} - Substring match (wildcard)
}


The query string should follow the pattern: \code{KEY[OPERATOR]VALUE}\cr
Supported operators:
\itemize{
  \item \code{=} - Equal to
  \item \code{>} - Greater than
  \item \code{<} - Less than
  \item \code{>=} - Greater than or equal to
  \item \code{<=} - Less than or equal to
  \item \code{*} - Substring match (wildcard)
}
}

\section{Available Output Formats}{

\describe{
  \item{\code{"df"}}{Returns results as a data.frame}
  \item{\code{"json"}}{Returns results as JSON string}
  \item{\code{"dict"}}{Returns results as named list (default)}
  \item{\code{"table"}}{Returns results in pretty-printed table format}
}
#'


\describe{
  \item{\code{"df"}}{Returns results as a data.frame}
  \item{\code{"json"}}{Returns results as JSON string}
  \item{\code{"dict"}}{Returns results as named list (default)}
  \item{\code{"table"}}{Returns results in pretty-printed table format}
}
#'
}

\examples{
\dontrun{
# Basic usage
search_collections_query(self, 'collection_name*lake')

# Multiple conditions
search_collections_query(self,
                        'collection_name*lake',
                        'inserted_by=user1@gmail.com',
                        'inserted_at>1747934722',
                        'public=TRUE',
                        output_format='table')
}

\dontrun{
# Basic usage
search_files_query(self, 'file_name*lake')

# Multiple conditions
search_files_query(self,
                        'file_name*lake',
                        'inserted_by=user1@gmail.com',
                        'inserted_at>1747934722',
                        'public=TRUE',
                        output_format='table')
}

}
\seealso{
\code{\link{parse_query_args}} for the query parsing implementation

\code{\link{parse_query_args}} for the query parsing implementation
}
